requirements.txt
torch==1.13.0
torchvision==0.14.0
torchaudio==0.13.0
Flask==2.2.2
Flask-Cors==3.0.10
gunicorn==20.1.0
nltk==3.7
google-api-python-client==2.66.0
transformers==4.18.0
spacy==3.4.4
opencv-python
keytotext==2.3.2
tensorflow-macos==2.12.0 
protobuf==3.20
